+++ scripts/functions.js
@@ -828,9 +828,10 @@ const blacklisted = (blacklist, url) => {
 }
 
 const dataTile = (subjects, elem, value) => {
+	console.log(subjects, elem);
 	const subject = subjects.find(subject => 
 		subject["characters"] === value ||
-		subject?.character_images?.find(image => image["url"] == elem.querySelector("img")?.src));
+		subject?.character_images?.find(image => image["url"] == elem.querySelector("image")?.getAttribute("src")));
 		
 	if (subject) {
 		const meaning = subject["meanings"][0];